internal/requirements.txt
flask
pymysql
fake-factory
Pillow
pytesseract
